use double type

+++ gnnflow/distributed/dispatcher.py
@@ -93,7 +93,7 @@ class Dispatcher:
             kvstore_rank = partition_id * self._local_world_size
             if edge_feats is not None:
                 keys = edges[3]
-                features = torch.from_numpy(edge_feats[keys.numpy()])
+                features = torch.from_numpy(edge_feats[keys.numpy()]).double()
                 futures.append(rpc.rpc_async("worker%d" % kvstore_rank, graph_services.push_tensors,
                                              args=(keys, features, 'edge')))
